requirements.txt
django==4.0.2
python-dotenv==0.19.2
psycopg2==2.9.3
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
django-jutil==3.10.4